misc cleanup
