+++ tool/test/org/antlr/v4/test/TestUtils.java
@@ -80,7 +80,9 @@ public class TestUtils {
     public static List<String> getTokenTypes(LexerGrammar lg,
                                              ATN atn,
                                              CharStream input) {
-        LexerATNSimulator interp = new LexerATNSimulator(atn, new DFA[]{new DFA(atn.modeToStartState.get(Lexer.DEFAULT_MODE))}, null);
+        LexerATNSimulator interp = new LexerATNSimulator(atn,
+                                                         new DFA[]{new DFA(atn.modeToStartState.get(Lexer.DEFAULT_MODE))},
+                                                         null);
         List<String> tokenTypes = new ArrayList<String>();
         int ttype;
         boolean hitEOF = false;
@@ -104,43 +106,6 @@ public class TestUtils {
         return tokenTypes;
     }
 
-    public static List<ANTLRMessage> checkRuleDFA(String gtext, String ruleName, String expecting)
-            throws Exception {
-        ErrorQueue equeue = new ErrorQueue();
-        Grammar g = new Grammar(gtext, equeue);
-        ATN atn = createATN(g, false);
-        ATNState s = atn.ruleToStartState[g.getRule(ruleName).index];
-        if (s == null) {
-            System.err.println("no such rule: " + ruleName);
-            return null;
-        }
-        ATNState t = s.transition(0).target;
-        if (!(t instanceof DecisionState)) {
-            System.out.println(ruleName + " has no decision");
-            return null;
-        }
-        DecisionState blk = (DecisionState) t;
-        checkRuleDFA(g, blk, expecting);
-        return equeue.all;
-    }
-
-    public static List<ANTLRMessage> checkRuleDFA(String gtext, int decision, String expecting)
-            throws Exception {
-        ErrorQueue equeue = new ErrorQueue();
-        Grammar g = new Grammar(gtext, equeue);
-        ATN atn = createATN(g, false);
-        DecisionState blk = atn.decisionToState.get(decision);
-        checkRuleDFA(g, blk, expecting);
-        return equeue.all;
-    }
-
-    public static void checkRuleDFA(Grammar g, DecisionState blk, String expecting)
-            throws Exception {
-        DFA dfa = createDFA(g, blk);
-        String result = null;
-        if (dfa != null) result = dfa.toString();
-        assertEquals(expecting, result);
-    }
 
     public static void checkRuleATN(Grammar g, String ruleName, String expecting) {
         DOTGenerator dot = new DOTGenerator(g);
@@ -155,56 +120,6 @@ public class TestUtils {
         assertEquals(expecting, result);
     }
 
-    public static List<ANTLRMessage> checkLexerDFA(String gtext, String expecting)
-            throws Exception {
-        return checkLexerDFA(gtext, LexerGrammar.DEFAULT_MODE_NAME, expecting);
-    }
-
-    public static List<ANTLRMessage> checkLexerDFA(String gtext, String modeName, String expecting)
-            throws Exception {
-        ErrorQueue equeue = new ErrorQueue();
-        LexerGrammar g = new LexerGrammar(gtext, equeue);
-        g.atn = createATN(g, false);
-//		LexerATNToDFAConverter conv = new LexerATNToDFAConverter(g);
-//		DFA dfa = conv.createDFA(modeName);
-//		g.setLookaheadDFA(0, dfa); // only one decision to worry about
-//
-//		String result = null;
-//		if ( dfa!=null ) result = dfa.toString();
-//		assertEquals(expecting, result);
-//
-//		return equeue.all;
-        return null;
-    }
-
-    public static DFA createDFA(Grammar g, DecisionState s) {
-//		PredictionDFAFactory conv = new PredictionDFAFactory(g, s);
-//		DFA dfa = conv.createDFA();
-//		conv.issueAmbiguityWarnings();
-//		System.out.print("DFA="+dfa);
-//		return dfa;
-        return null;
-    }
-
-    public static void checkGrammarSemanticsError(ErrorQueue equeue,
-                                                  GrammarSemanticsMessage expectedMessage)
-            throws Exception {
-        ANTLRMessage foundMsg = null;
-        for (int i = 0; i < equeue.errors.size(); i++) {
-            ANTLRMessage m = equeue.errors.get(i);
-            if (m.getErrorType() == expectedMessage.getErrorType()) {
-                foundMsg = m;
-            }
-        }
-        assertNotNull("no error; " + expectedMessage.getErrorType() + " expected", foundMsg);
-        assertTrue("error is not a GrammarSemanticsMessage",
-                foundMsg instanceof GrammarSemanticsMessage);
-        assertEquals(Arrays.toString(expectedMessage.getArgs()), Arrays.toString(foundMsg.getArgs()));
-        if (equeue.size() != 1) {
-            System.err.println(equeue);
-        }
-    }
-
     public static void checkGrammarSemanticsWarning(ErrorQueue equeue,
                                                     GrammarSemanticsMessage expectedMessage)
             throws Exception {
@@ -217,75 +132,18 @@ public class TestUtils {
         }
         assertNotNull("no error; " + expectedMessage.getErrorType() + " expected", foundMsg);
         assertTrue("error is not a GrammarSemanticsMessage",
-                foundMsg instanceof GrammarSemanticsMessage);
+                   foundMsg instanceof GrammarSemanticsMessage);
         assertEquals(Arrays.toString(expectedMessage.getArgs()), Arrays.toString(foundMsg.getArgs()));
         if (equeue.size() != 1) {
             System.err.println(equeue);
         }
     }
 
-    public static void checkError(ErrorQueue equeue,
-                                  ANTLRMessage expectedMessage)
-            throws Exception {
-        //System.out.println("errors="+equeue);
-        ANTLRMessage foundMsg = null;
-        for (int i = 0; i < equeue.errors.size(); i++) {
-            ANTLRMessage m = equeue.errors.get(i);
-            if (m.getErrorType() == expectedMessage.getErrorType()) {
-                foundMsg = m;
-            }
-        }
-        assertTrue("no error; " + expectedMessage.getErrorType() + " expected", !equeue.errors.isEmpty());
-        assertTrue("too many errors; " + equeue.errors, equeue.errors.size() <= 1);
-        assertNotNull("couldn't find expected error: " + expectedMessage.getErrorType(), foundMsg);
-        /*
-        assertTrue("error is not a GrammarSemanticsMessage",
-				   foundMsg instanceof GrammarSemanticsMessage);
-		 */
-        assertArrayEquals(expectedMessage.getArgs(), foundMsg.getArgs());
-    }
-
-    /**
-     * When looking at a result set that consists of a Map/HashTable
-     * we cannot rely on the output order, as the hashing algorithm or other aspects
-     * of the implementation may be different on differnt JDKs or platforms. Hence
-     * we take the Map, convert the keys to a List, sort them and Stringify the Map, which is a
-     * bit of a hack, but guarantees that we get the same order on all systems. We assume that
-     * the keys are strings.
-     *
-     * @param m The Map that contains keys we wish to return in sorted order
-     * @return A string that represents all the keys in sorted order.
-     */
-    public static <K, V> String sortMapToString(Map<K, V> m) {
-        // Pass in crap, and get nothing back
-        //
-        if (m == null) {
-            return null;
-        }
-
-        System.out.println("Map toString looks like: " + m.toString());
-
-        // Sort the keys in the Map
-        //
-        TreeMap<K, V> nset = new TreeMap<K, V>(m);
-
-        System.out.println("Tree map looks like: " + nset.toString());
-        return nset.toString();
-    }
 
     public static List<String> realElements(List<String> elements) {
         return elements.subList(Token.MIN_USER_TOKEN_TYPE, elements.size());
     }
 
-    /**
-     * Sort a list
-     */
-    public static <T extends Comparable<? super T>> List<T> sort(List<T> data) {
-        List<T> dup = new ArrayList<T>();
-        dup.addAll(data);
-        Collections.sort(dup);
-        return dup;
-    }
 
     /**
      * Return map sorted by key
@@ -301,22 +159,6 @@ public class TestUtils {
         return dup;
     }
 
-    public static <T> List<? super T> filter(List<? extends T> input, Class<T> include) {
-        List<? super T> filtered = new ArrayList<T>();
-        for (T t : input) {
-            if (t.getClass() == include) filtered.add(t);
-        }
-        return filtered;
-    }
-
-
-    public static String getFirstLineOfException(String stackTrace) {
-
-        String[] lines = stackTrace.split("\n");
-        String prefix = "Exception in thread \"main\" ";
-        return lines[0].substring(prefix.length(), lines[0].length());
-    }
-
     public static Matcher<List<Diagnostic<? extends JavaFileObject>>> containsNoErrors() {
         return CHECK_LIST_FOR_ERRORS;
     }
@@ -324,9 +166,6 @@ public class TestUtils {
     public static Matcher<DiagnosticCollector<JavaFileObject>> hasNoErrors() {
         return CHECK_DIAGNOSTICS_FOR_ERRORS;
     }
-
-
-
 
 
     static final Matcher<List<Diagnostic<? extends JavaFileObject>>> CHECK_LIST_FOR_ERRORS =
@@ -342,7 +181,8 @@ public class TestUtils {
                 }
 
                 @Override
-                protected void describeMismatchSafely(List<Diagnostic<? extends JavaFileObject>> item, org.hamcrest.Description mismatchDescription) {
+                protected void describeMismatchSafely(List<Diagnostic<? extends JavaFileObject>> item,
+                                                      org.hamcrest.Description mismatchDescription) {
                     for (Diagnostic<? extends JavaFileObject> diagnostic : item) {
                         if (diagnostic.getKind() == Diagnostic.Kind.ERROR) {
                             mismatchDescription
@@ -362,8 +202,8 @@ public class TestUtils {
                     }
 
                     mismatchDescription.appendText("\n all diagnostics:\n")
-                            .appendValueList("[", "\n\t", "]", item)
-                            .appendText("\n");
+                                       .appendValueList("[", "\n\t", "]", item)
+                                       .appendText("\n");
                 }
             };
     static final Matcher<DiagnosticCollector<JavaFileObject>> CHECK_DIAGNOSTICS_FOR_ERRORS =
