+++ tool/test/org/antlr/v4/test/impl/AntlrTestDelegate.java
@@ -41,7 +41,7 @@ public interface AntlrTestDelegate {
                                        boolean defaultListener,
                                        String... extraOptions);
 
-    String getGenPath();
+    String getWorkingDir();
 
     String getStdErrDuringParse();
